refactor to split library from node script + included clean indentation of 4 spaces
