+++ bin/mwoffliner.script.js
@@ -1,0 +1,28 @@
+#!/bin/sh
+":" //# -*- mode: js -*-; exec /usr/bin/env node --max_inlined_source_size=100000 --max-old-space-size=9000 --stack-size=42000 "$0" "$@"
+
+"use strict";
+
+const yargs = require( 'yargs' );
+const mwofflinerLib = require('../lib/mwoffliner.lib.js')
+
+/************************************/
+/* Command Parsing ******************/
+/************************************/
+
+const parametersList = mwofflinerLib.getParametersList()
+
+const argv = yargs
+    .usage(`Create a fancy HTML dump of a Mediawiki instance in a directory
+Usage: $0'
+Example, as a node script:
+    node mwoffliner.js --mwUrl=https://en.wikipedia.org/ --adminEmail=foo@bar.net
+Or, as a npm script: '
+    npm run mwoffliner -- --mwUrl=https://en.wikipedia.org/ --adminEmail=foo@bar.net`)
+    .require(parametersList.filter(param => param.required).map(param => param.name))
+
+parametersList.forEach(param => argv.describe(param.name, param.description))
+
+argv.strict().argv
+
+mwofflinerLib.execute(argv)
