+++ index.js
@@ -1,0 +1,2 @@
+var mwoffliner = require('./lib/mwoffliner.lib.js')
+module.exports = mwoffliner
